Cascade/Exp/CascadeProjects/windsurf-project/requirements.txt
flask
transformers
huggingface_hub
python-dotenv
gunicorn
flask-cors
torch
